Combined game into starting screen. Created 2nd and 3rd game mode.
