+++ app/src/main/AndroidManifest.xml
@@ -17,8 +17,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".GameOne" android:noHistory="true"/>
-        <activity android:name=".GameTwo" android:noHistory="true"/>
     </application>
 
 </manifest>
